requirements.txt
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
sounddevice==0.4.6
openai-whisper==20231117
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
torch>=2.0.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
openai-whisper==20231117
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
seaborn==0.12.2
torch==2.0.1
spacy
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0